Decompile sef/sc/sv continuation (23 functions)
Script-interpreter register ops (scRADD/RSUB/RMUL/R_AND/R_OR/R_XOR/
R_NEG/R_NOT/RSET/RINC/RDEC), effect draw/load, and image-mapper helpers.

New decoding trap worth remembering: `lui a0,0x400` + `and` is a mask of
0x04000000, not 0x400 -- lui sets bits 31:16. A previous derivation read
GameLoopState.nFlags & 0x400 and was wrong. Any lui+and pair against a
small-looking constant should be re-decoded as imm << 16 before assuming
it is a byte-sized flag; the giveaway is that a genuine small mask would
use andi and need no lui at all.

+++ src/sef.c
@@ -144,7 +144,7 @@ extern float D_004D82D8;
  * draw, scale by 1/32767. */
 float sefRandf(void) {
     sefRandSeed = sefRandSeed * 0x41C64E6D + 12345;
-    return (float)(((unsigned int)sefRandSeed >> 16) & 0x7FFF) * D_004D82D8;
+    return (float)(int)(((unsigned int)sefRandSeed >> 16) & 0x7FFF) * D_004D82D8;
 }
 
 extern int sefCnvEtEffectNo(int a);
